Implement ban API and work more on Minecraft/Dedicated server impls (#54)
* Split mixin launcher into separate coremod and get mixins to *load*

This still crashes the server with a NoClassDefFoundError. The cause of this has yet to be identified, but people on Discord say it's usually due to the class being loaded or referenced when it shouldn't be. However, I've looked and it doesn't appear we're doing so at all. Will need further investigation.

* Fix NoClassDefFoundError and get Magma to run

* Add BungeeCord repo