Fix plugin loading (#56)
* Fix plugin loading

This requires us to relocate commons-lang to a separate package for now, which should hopefully only server as a temporary solution. Plugins sometimes rely on this library and with it relocated, those plugins will not work for the time being. This is only a *temporary* solution until the root cause of the issue can be found.

The strange part is during the plugin loading process, a NoClassDefFoundError is thrown for the Validate class in commons-lang, despite it being inside of the jar itself. This is likely due to a classloader issue somewhere, but where that is is still unknown.

* Minor fixes to get farther along the plugin load process